added string interning classes, not yet integrated

+++ Code/SequenceController/src/CommandSource.h
@@ -70,7 +70,7 @@ public:
 			pEnd++;
 		}
 
-		_command = Command(pCurrent, pEnd - pCurrent, commandNumber);
+		_command = Command(pCurrent, (int) (pEnd - pCurrent), commandNumber);
 
 		return &_command;
 	}
